% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tritopatch-package.R
\docType{package}
\name{tritopatch-package}
\alias{tritopatch}
\alias{tritopatch-package}
\title{tritopatch: Tritopatch}
\description{
This project contains custom functions with roxygen2-generated documentation.
}
\keyword{internal}
